+ quarters string fixes + Aero attack values and ranges for clan RACs

+++ megamek/src/megamek/common/PillionSeatCargoBay.java
@@ -72,7 +72,7 @@ public final class PillionSeatCargoBay extends Bay {
     @Override
     public String getUnusedString() {
         StringBuffer returnString = new StringBuffer("Passenger Pillion Seats - ");
-        returnString.append(currentSpace);
+        returnString.append((int)currentSpace);
         return returnString.toString();
     }
